fix room description in node modal header
