+++ packages/client/src/network/shapes/Room/index.ts
@@ -1,9 +1,10 @@
+export { NullRoom } from './constants';
 export { getExitsFor as getExitsForRoom } from './exit';
 export { getAllRooms, getRoomByIndex } from './functions';
 export { getGates } from './gate';
 export { getRoomsX } from './getters';
 export { queryByIndex as queryRoomByIndex, query as queryRooms } from './queries';
-export { NullRoom, getRoom } from './types';
+export { getRoom } from './types';
 
 export type { Exit } from './exit';
 export type { QueryOptions } from './queries';
